# 06-jirafeitor/server/models/postgres/User.go
package postgres

type PostgresUser struct {
	ID       int
	Username string
	Password string
}

type PostgresUserPost struct {
	Username string
	Password string
}

type PostgresUserInfo struct {
	ID       int
	Username string
}